Use /bin/kill on Cygwin.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@60187 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/common/check/check_exec_guard.sh
@@ -25,14 +25,20 @@ timeout=$1
 pid=$2
 sleep_time=5
 
+# Use different kill on Unix and Cygwin
+case `uname -s` in
+   CYGWIN* ) cygwin=true  ; kill='/bin/kill -f' ;;
+   *)        cygwin=false ; kill='kill' ;;
+esac
+
 # Wait 
 while [ $timeout -gt 0 ]; do
-  kill -0 $pid > /dev/null 2>&1  ||  exit 0
-  if [ $timeout -lt  $sleep_time ]; then
-    sleep_time=$timeout
-  fi
-  timeout="`expr $timeout - $sleep_time`"
-  sleep $sleep_time >/dev/null 2>&1
+   kill -0 $pid > /dev/null 2>&1  ||  exit 0
+   if [ $timeout -lt  $sleep_time ]; then
+      sleep_time=$timeout
+   fi
+   timeout="`expr $timeout - $sleep_time`"
+   sleep $sleep_time >/dev/null 2>&1
 done
 
 # Time out, kill the process
@@ -40,8 +46,10 @@ echo
 echo "Maximum execution time of $1 seconds is exceeded"
 echo
 
-kill $pid > /dev/null 2>&1
-sleep $sleep_time >/dev/null 2>&1
-kill -9 $pid > /dev/null 2>&1
+$kill $pid > /dev/null 2>&1
+if ! $cygwin; then
+   sleep $sleep_time >/dev/null 2>&1
+   kill -9 $pid > /dev/null 2>&1
+fi
 
 exit 1
